requirements.txt
# Core dependencies (successfully installed)
opencv-python>=4.8.0
flask>=2.3.0
flask-socketio>=5.3.0
pywebview>=4.4.0
numpy>=1.24.0
python-socketio>=5.8.0

# MediaPipe (optional - fallback available)
# Note: MediaPipe not available for Python 3.13 as of 2024
# mediapipe>=0.10.0